# Performance Improvement 1 - Reduce number of pure calls.