`is_ignored_block_opening(config, item)` and `is_safe_closing_tag(config, item)` are each called multiple times in indent.py with the same parameters. This can be deduplicated.

(Note I've changed the block `with executor_cls(max_workers=worker_count) as exe:` in `__init__.py` to be serial to make the profiler work.
```python
for this_file in file_list:
  file_errors.append(process(config, this_file))
  # if temp_file block etc. is removed
```
# Before:

![image](https://github.com/user-attachments/assets/71f4ba63-8415-463f-9bfc-8a87028aad41)

is_ignored_block_opening (With children): 12.4% of all the time, 61562x calls
is_safe_closing_tag (With children): 6.24% of all the time, 33139x calls

Concrete data:
```
Netbox with serial execution (Patch above): 9.7s
Netbox with unchanged __init__.py              :    4s
OpenEDX platform                                           : 33s
```

# After
![image](https://github.com/user-attachments/assets/f1ef6531-6344-4a8e-bf13-885781937256)

is_ignored_block_opening (With children): 4.73% of all the time, 20622x calls
is_safe_closing_tag (With children): 4.61% of all the time, 20941x calls

Concrete data:
```
Netbox with serial execution (Patch above): 8.3s
Netbox with unchanged __init__.py              : 3.5s
OpenEDX platform                                           : 26s
```